requirements.txt
couchbase==4.1.2
PyYAML==6.0
